#25 fix: 리뷰 피드백 반영 - TimeDescription/getTagName 중복 코드 추출
- TimeDescription을 공유 컴포넌트로 추출 (src/components/TimeDescription.tsx)
- getTagName을 커스텀 훅으로 추출 (src/hooks/useTagName.ts)
- place 필드는 모델에 없어 스코프 외
